feat: activation explorer site, CI infrastructure, 72 activation fixes
- add Next.js static site with family nav, property filters, KaTeX formulas, SVG plots, Pyodide sliders, shareable permalinks
- add GitHub Pages deploy workflow
- add pre-commit hooks and lint CI (black, isort, flake8, doc8, docstring checker)
- add numerical auto-tagger for overflow-safe/monotonic/bounded/smooth properties
- add SVG plot generator and data.json build pipeline
- fix 72 activation functions: overflow clamps, scalar tensor params, boolean mask indexing, shape mismatches, docstring corrections
- add docs RST pages and tests for 415 activations
- update ROADMAP with all completed work

+++ torch_activation/adaptive/melu.py
@@ -58,12 +58,10 @@ class MeLU(BaseActivation):
 
     def _forward(self, x) -> Tensor:
         # PReLU part
-        prelu_out = F.prelu(x, self.prelu_weight)
-        
-        # Sum part
+        prelu_out = F.prelu(x, self.prelu_weight.to(x.dtype))
+
         sum_part = torch.zeros_like(x)
         for j in range(self.k-1):
-            # Calculate phi_b_j,c_j(z_i)
             phi = torch.clamp(self.c[j] - torch.abs(x - self.b[j]), min=0.0)
             sum_part += self.a[j] * phi
         
@@ -169,9 +167,8 @@ class GaLU(BaseActivation):
 
     def _forward(self, x) -> Tensor:
         # PReLU part
-        prelu_out = F.prelu(x, self.prelu_weight)
-        
-        # Sum part
+        prelu_out = F.prelu(x, self.prelu_weight.to(x.dtype))
+
         sum_part = torch.zeros_like(x)
         for j in range(self.k-1):
             # Calculate phi_b_j,c_j(z_i) for GaLU
